Fixed some test and code climate errors
