+++ app/react/Templates/specs/utils/renderConnected.tsx
@@ -3,6 +3,7 @@ import { mount, ReactWrapper, shallow } from 'enzyme';
 import configureStore, { MockStore, MockStoreCreator } from 'redux-mock-store';
 import { ConnectedComponent, Provider } from 'react-redux';
 import thunk from 'redux-thunk';
+import Immutable from 'immutable';
 
 const middlewares = [thunk];
 const mockStoreCreator: MockStoreCreator<object> = configureStore<object>(middlewares);
@@ -36,9 +37,21 @@ const renderConnected = (
 const renderConnectedMount = (
   Component: ConnectedComponent<any, any> | ((props: any) => any),
   state: any = {},
-  props: any = {}
+  props: any = {},
+  useDefaultTranslationState = false
 ): ReactWrapper<React.Component['props'], React.Component['state'], React.Component> => {
-  const store = mockStoreCreator(state);
+  const defaultState = {
+    locale: 'en',
+    inlineEdit: Immutable.fromJS({ inlineEdit: true }),
+    translations: Immutable.fromJS([
+      {
+        locale: 'en',
+        contexts: [],
+      },
+    ]),
+  };
+  const stateToMap = { ...defaultState, state };
+  const store = mockStoreCreator(useDefaultTranslationState ? stateToMap : state);
   return mount(
     <Provider store={store}>
       <Component {...props} />
